changed types on all functions, added devision by zero exception for file devision.cpp

+++ subtraction.cpp
@@ -1,4 +1,4 @@
 #include <subtraction.h>
-int subtraction (int a, int b) {
+float subtraction (float a, float b) {
     return a-b;
 }
